tests/wpt/web-platform-tests/bluetooth/requestDevice/request-from-sandboxed-iframe.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<body>
<script>
'use strict';
const test_desc = 'Request device from a unique origin. ' +
    'Should reject with SecurityError.';
const expected = 'SecurityError: requestDevice() called from cross-origin ' +
    'iframe.';

let iframe = document.createElement('iframe');

bluetooth_test(
    () => getConnectedHealthThermometerDevice()
              // 1. Load the iframe.
              .then(() => new Promise(resolve => {
                      iframe.sandbox.add('allow-scripts');
                      iframe.src =
                          '/bluetooth/resources/health-thermometer-iframe.html';
                      document.body.appendChild(iframe);
                      iframe.addEventListener('load', resolve);
                    }))
              // 2. Request the device from the iframe.
              .then(() => new Promise(resolve => {
                      callWithTrustedClick(() => {
                        iframe.contentWindow.postMessage(
                            {type: 'RequestDevice'}, '*');
                      });

                      window.onmessage = messageEvent => {
                        assert_equals(messageEvent.data, expected);
                        resolve();
                      }
                    })),
    test_desc);
</script>
</body>
